+++ CHANGELOG.md
@@ -6,6 +6,8 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 ## [Unreleased]
+
+## [0.0.75] 2023-08-23
 ### Added
 - Added support for passing in the granularity when running queries
 - Added fetching the datasets from the backend when prompting to select a dataset